Restore italic and bold to Emphasis markup (#4)
This patch restores the italics and bold highlights to Emphasis and
Strong Emphasis inline markups.

We also append the default syntax highlight colours for Underline and
Special regions, in order to match the previous version.

This presumes that Underlined and Special (actually cterm 13 and 1) are
safe foreground colours.  We can revisit this later if it's a problem.