Leading whitespace on references
Previously, footnotes and reference highlighting did not check for
whitespace, so that this appeared valid, but would not produce a
footnote (as required by the spec):

    foo[1]_

We now require whitespace between the text and the footnote to correctly
indicate and highlight the footnote.

Thanks (as always) to Antony Lee (anntzer) for finding this issue.