LiteralBlock: replacing match start with lookback
LiteralBlock matching was changed significantly to begin its match at
the start of the line, rather than at the directive marker (::).

This has managed to resolve all previous issues with literalblock and is
currently passing all of the current limited set of tests.

Big debt of gratitude to Rich at the Vi[m] Stack Exchange:

    https://vi.stackexchange.com/questions/18732/syntax-highlighting-a-region-inside-the-start-match-of-another-region

Regex explanation is below.

----

This method stores the leading whitespace to \z1 :

    ^\z(\s*\)

and then does a lookbehind for the text leading up to the directive
header termination, using the generic character glob .* :

    \(^\z(\s*\).*\)\@<=::

followed by a mandatory blank line:

    start=\(^\z(\s*\).*\)\@<=::\n\s*\n

We now skip all blank lines which don't match the whitespace (\z1):

    skip='^\s*$'

and terminate once we reach a line which whitespace that is not longer
than \z1 (i.e. \z1\s\+):

    end='^\(\z1\s\+\)\@!