Single char escape for multichar delimiters
When a delimiter had more than one character (e.g. ** or ``), the syntax
file would try to escape the full sequence with a slash (\** or \``)
which is not correct.

We now only try to escape the first character of a multicharacter
delimiter, under the assumption that it is always a repeated character.