Language name part of "code" directive
If a language is specified in the `..code::` directive but is not in
rst_syntax_code_list, then it would not highlight the code block, which
would cause other problems.

This problem was introduced when fields were formally added as part of
the description of the code block.

This patch supports a single (space-delimited) code name along with the
directive.

A very minimal "test" file for code blocks has been added.