Experimental rewrite of directive handling
I am testing out a new approach to directives.  It clobbers the old
heavly nested model, with lots of clusters and nextgroups, and instead
rewrites every explicit markup block on its own.

The current approach does not appear to provide a way to track leading
whitespace, whereas this is not an issue with independent regions.

(Specifically, I can't define "\z" groups in one match and use them in
another.)

This has the potential to break *E V E R Y T H I N G*, so I need to test
this carefully...