Comment rewrite
Comment highlights were changed rather dramatically.  We now start with
the .. token and say anything after this is a comment.

Then we layer the other explicit block types afterwards, so that those
special cases override the comment notation.

As correct as this this, it will be "wrong" for while until the other
types are added.